+++ src/main/java/com/dfoff/demo/Domain/JsonDtos/CharacterAbilityDto.java
@@ -1,4 +1,4 @@
-package com.dfoff.demo.Domain.ForCharacter;
+package com.dfoff.demo.Domain.JsonDtos;
 
 
 import com.dfoff.demo.Domain.CharacterEntity;
