+++ src/main/java/com/dfoff/demo/Domain/JsonDtos/CharacterEquipmentJsonDto.java
@@ -1,5 +1,5 @@
 
-package com.dfoff.demo.Domain.ForCharacter;
+package com.dfoff.demo.Domain.JsonDtos;
 
 import java.io.Serializable;
 import java.util.ArrayList;
