+++ src/main/java/com/dfoff/demo/Domain/JsonDtos/CharacterSkillDetailJsonDto.java
@@ -1,5 +1,5 @@
 
-package com.dfoff.demo.Domain.ForCharacter;
+package com.dfoff.demo.Domain.JsonDtos;
 
 import java.io.Serializable;
 import java.util.List;
